edit event API and delete event API, modified add favorite API so that user can not add his own event to favorite

+++ controllers/favorite.controller.js
@@ -5,44 +5,48 @@ const mongoose = require("mongoose");
 
 // create or delete a favorite event
 const addToFavorite = async (req, res) => {
-  const {id} = req.params;
+  const { id } = req.params;
   let userId = req.decoded._id;
 
-  if(!id){
-    throw new CustomError("Event Id is required", 403)
+  if (!id) {
+    throw new CustomError("Event Id is required", 403);
   }
 
   // convert string ids to ObjectId id
   const eventObjectId = new mongoose.Types.ObjectId(id);
   const userObjectId = new mongoose.Types.ObjectId(userId);
 
-  const existingFavorite = await Favorite.findOne({eventId: eventObjectId, userId: userObjectId});
+  // user cannot add his own event to favorite
+  const isSameUser = await Event.findOne({ _id: id, userId });
+  if (isSameUser) throw new CustomError("You can not add your own evnet to favorite", 403);
 
-  if(existingFavorite){
-    const removed = await Favorite.deleteOne({_id: existingFavorite._id});
-    return res.status(200).json({status: "success", data: removed})
+  const existingFavorite = await Favorite.findOne({ eventId: eventObjectId, userId: userObjectId });
+
+  if (existingFavorite) {
+    const removed = await Favorite.deleteOne({ _id: existingFavorite._id });
+    return res.status(200).json({ status: "success", data: removed })
   }
 
-  const result = await Favorite.create({eventId: eventObjectId, userId: userObjectId});
+  const result = await Favorite.create({ eventId: eventObjectId, userId: userObjectId });
 
-  return res.status(201).json({status: "success", data: result})
+  return res.status(201).json({ status: "success", data: result })
 }
 
 /// get all favorite events
 const getFavorites = async (req, res) => {
   const userId = req.decoded._id;
 
-  const {fields, sort} = req.query;
+  const { fields, sort } = req.query;
 
   // selective fields
   //NOTE: populate is like joins in SQL
-  let result = Favorite.find({userId}).populate({path: "eventId", select: fields ? fields.split(",").join(" ") : ""});
+  let result = Favorite.find({ userId }).populate({ path: "eventId", select: fields ? fields.split(",").join(" ") : "" });
 
   // sorting
-  if(sort){
+  if (sort) {
     const sortBy = sort.split(",").join(" ");
     result = result.sort(sortBy);
-  }else{
+  } else {
     result = result.sort("createdAt");
   }
 
@@ -55,30 +59,30 @@ const getFavorites = async (req, res) => {
 
   finalResult = await result;
 
-  if(!finalResult){
+  if (!finalResult) {
     throw CustomError("Event not found", 404);
   }
 
   const events = finalResult.map(fav => fav.eventId);
-  return res.status(200).json({status: "success", data: events});
+  return res.status(200).json({ status: "success", data: events });
 }
 
 // get single favorite event my user id and event id
 const getFavoritesByEventIdAndUserId = async (req, res) => {
-  const {id} = req.params;
+  const { id } = req.params;
   const userId = req.decoded._id;
 
-  if(!id){
+  if (!id) {
     throw new CustomError("Event Id is required", 403);
   }
 
-  const result = await Favorite.findOne({eventId: id, userId});
-  
-  if(!result){
+  const result = await Favorite.findOne({ eventId: id, userId });
+
+  if (!result) {
     throw new CustomError("Event doesn't exists", 404);
   }
 
-  return res.status(200).json({status: "success", data: result});
+  return res.status(200).json({ status: "success", data: result });
 }
 
 module.exports = {
